contrib/netsurf/libhubbub/test/data/tree-construction/after-after-frameset.dat

#data
<!DOCTYPE html><html><frameset></frameset></html><!--x-->
#errors
#comments
"A comment token"
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <frameset>
| <!-- x -->

#data
<!DOCTYPE html><html><frameset></frameset></html><!DOCTYPE foo>
#errors
Unexpected DOCTYPE: <!DOCTYPE foo>
#comments
"A DOCTYPE token"
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <frameset>

#data
<!DOCTYPE html><html><frameset></frameset></html>

#errors
#comments
Whitespace character token
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <frameset>
|   "
"

#data
<!DOCTYPE html><frameset></frameset></html><html bgcolor="red">
#errors
Second <html> tag not allowed.
#comments
A start tag whose tag name is "html"
#document
| <!DOCTYPE html>
| <html>
|   bgcolor="red"
|   <head>
|   <frameset>

#data
<!DOCTYPE html><html><frameset></frameset></html>
#errors
#comments
"An end-of-file token"
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <frameset>

#data
<!DOCTYPE html><html><frameset></frameset></html><noframes>
#errors
End tag after </html>.
</frameset> when the current node is the root "html" element.
#comments
A start tag whose tag name is "noframes"
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <frameset>
|   <noframes>

#data
<!DOCTYPE html><html><frameset></frameset></html>xx
#errors
Text after </html> tag.
Text in frameset.
#comments
"Anything else" (character token)
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <frameset>

#data
<!DOCTYPE html><html><frameset></frameset></html><table>
#errors
Tag after </html>.
<table> not allowed in <frameset>.
#comments
"Anything else" (start tag)
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <frameset>

#data
<!DOCTYPE html><html><frameset></frameset></html></frameset>
#errors
End tag after </html>.
</frameset> when the current node is the root "html" element.
#comments
"Anything else" (end tag)
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <frameset>
